[chore] 패키지 및 파일 분리
map, main 패키지 파일 분리
