+++ app/src/main/java/com/squirtles/musicroad/main/MainActivity.kt
@@ -12,6 +12,8 @@ import androidx.core.content.PermissionChecker
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
 import androidx.navigation.compose.rememberNavController
 import com.squirtles.musicroad.R
+import com.squirtles.musicroad.main.navigations.MainNavGraph
+import com.squirtles.musicroad.main.navigations.MainNavigationActions
 import com.squirtles.musicroad.ui.theme.MusicRoadTheme
 import dagger.hilt.android.AndroidEntryPoint
