+++ app/src/main/java/com/squirtles/musicroad/map/components/InfoWindowCard.kt
@@ -1,4 +1,4 @@
-package com.squirtles.musicroad.map
+package com.squirtles.musicroad.map.components
 
 import android.content.res.Configuration.UI_MODE_NIGHT_YES
 import android.util.Size
